worker/requirements.txt
camunda_external_task_client_python3==4.5.0
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
requests==2.33.1
urllib3==2.6.3
